requirements.txt
# ── Core ──────────────────────────────────────────────────────────────────────
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
pydantic>=2.7.0
pydantic-settings>=2.2.0
python-multipart>=0.0.5

# ── ML ────────────────────────────────────────────────────────────────────────
scikit-learn>=1.4.0
xgboost>=2.0.0
lightgbm>=4.3.0
numpy>=1.26.0
pandas>=2.2.0
scipy>=1.13.0

# ── Hyperparameter tuning (optional) ─────────────────────────────────────────
scikit-optimize>=0.10.0   # Bayesian search

# ── Serialisation ─────────────────────────────────────────────────────────────
joblib>=1.4.0
PyYAML>=6.0.1

# ── Async / task queue ────────────────────────────────────────────────────────
celery[redis]>=5.3.0
redis>=5.0.0

# ── Visualisation ─────────────────────────────────────────────────────────────
plotly>=5.21.0
matplotlib>=3.8.0
seaborn>=0.13.0

# ── Streamlit ─────────────────────────────────────────────────────────────────
streamlit>=1.34.0

# ── Explainability ────────────────────────────────────────────────────────────
shap>=0.45.0

# ── ONNX export (optional) ────────────────────────────────────────────────────
# skl2onnx>=1.16.0
# onnx>=1.16.0

# ── Testing ───────────────────────────────────────────────────────────────────
pytest>=8.0.0
pytest-asyncio>=0.23.0
httpx>=0.27.0   # for FastAPI test client
